:recycle: Refactoring code
